LRCI-3961 Create a base open pull request event handler
